Big front-end refactor
Nothing changes from the user perspective, except for the
location of some things. In general, people shouldn't have to
touch as many files to understand a particular thing;
all of the Parameters specific to an object are now in that
object's file; e.g., NeuronTypeParam is in neurons.py.
This has cut down on some ugly imports within parameters,
and some other places that were a circular import waiting to happen.

Importantly, no functionality is changed in this commit.
All code is just straight moves, with some changes to
import statements to reflect their new locations.
