+++ nengo/tests/test_ensemble.py
@@ -1,3 +1,4 @@
+from collections import Counter
 import logging
 
 import numpy as np
@@ -5,6 +6,7 @@ import pytest
 
 import nengo
 import nengo.utils.numpy as npext
+from nengo.ensemble import NeuronTypeParam
 from nengo.utils.testing import Plotter, warns
 
 logger = logging.getLogger(__name__)
@@ -309,6 +311,39 @@ def test_gain_bias(Simulator, nl_nodirect):
     sim = Simulator(model)
     assert np.array_equal(gain, sim.data[a].gain)
     assert np.array_equal(bias, sim.data[a].bias)
+
+
+def test_neurontypeparam():
+    """NeuronTypeParam must be a neuron type."""
+    class Test(object):
+        ntp = NeuronTypeParam(default=None)
+
+    inst = Test()
+    inst.ntp = nengo.LIF()
+    assert isinstance(inst.ntp, nengo.LIF)
+    with pytest.raises(ValueError):
+        inst.ntp = 'a'
+
+
+def test_neurontypeparam_probeable():
+    """NeuronTypeParam can update a probeable list."""
+    class Test(object):
+        ntp = NeuronTypeParam(default=None, optional=True)
+        probeable = ['output']
+
+    inst = Test()
+    assert inst.probeable == ['output']
+    inst.ntp = nengo.LIF()
+    assert Counter(inst.probeable) == Counter(inst.ntp.probeable + ['output'])
+    # The first element is important,  as it's the default
+    assert inst.probeable[0] == 'output'
+    # Setting it again should result in the same list
+    inst.ntp = nengo.LIF()
+    assert Counter(inst.probeable) == Counter(inst.ntp.probeable + ['output'])
+    assert inst.probeable[0] == 'output'
+    # Unsetting it should clear the list appropriately
+    inst.ntp = None
+    assert inst.probeable == ['output']
 
 
 if __name__ == "__main__":
